refactor(FR #99): centralize Ireland utils to variants directory
Move Ireland-specific utility functions from src/utils/ to
src/config/variants/ireland/utils/ for better organization:

File moves:
- ireland-filter.ts → utils/filters.ts

Also:
- Create utils/index.ts for unified exports
- Update test file import path

Closes #99
