+++ src/config/variants/ireland/utils/filters.ts
