+++ src/config/variants/ireland/utils/index.ts
@@ -1,0 +1,14 @@
+/**
+ * Ireland Utils Exports
+ *
+ * Centralized exports for all Ireland-specific utility functions.
+ */
+
+export {
+  isInIreland,
+  containsIrelandKeywords,
+  isIrelandSource,
+  isIrelandRelated,
+  sortByIrelandRelevance,
+  type FilterableItem,
+} from './filters';
